Migrate ScopeProvider type to cursorless common
