+++ packages/cursorless-engine/src/index.ts
@@ -6,4 +6,3 @@ export * from "./core/StoredTargets";
 export * from "./typings/TreeSitter";
 export * from "./cursorlessEngine";
 export * from "./api/CursorlessEngineApi";
-export * from "./api/ScopeProvider";
