+++ packages/cursorless-engine/src/scopeProviders/getIterationScopeRanges.ts
@@ -1,6 +1,5 @@
-import { Range, TextEditor } from "@cursorless/common";
+import { IterationScopeRanges, Range, TextEditor } from "@cursorless/common";
 import { map } from "itertools";
-import { IterationScopeRanges } from "..";
 import { ModifierStage } from "../processTargets/PipelineStages.types";
 import { ScopeHandler } from "../processTargets/modifiers/scopeHandlers/scopeHandler.types";
 import { Target } from "../typings/target.types";
