+++ packages/cursorless-engine/src/scopeProviders/getTargetRanges.ts
@@ -1,6 +1,9 @@
-import { toCharacterRange, toLineRange } from "@cursorless/common";
+import {
+  TargetRanges,
+  toCharacterRange,
+  toLineRange,
+} from "@cursorless/common";
 import { Target } from "../typings/target.types";
-import { TargetRanges } from "../api/ScopeProvider";
 
 export function getTargetRanges(target: Target): TargetRanges {
   return {
