+++ packages/cursorless-vscode/src/ide/vscode/VSCodeScopeVisualizer/VscodeIterationScopeVisualizer.ts
@@ -1,7 +1,11 @@
-import { Disposable, TextEditor, toCharacterRange } from "@cursorless/common";
+import {
+  Disposable,
+  ScopeSupport,
+  TextEditor,
+  toCharacterRange,
+} from "@cursorless/common";
 import { VscodeTextEditorImpl } from "../VscodeTextEditorImpl";
 import { VscodeScopeVisualizer } from "./VscodeScopeVisualizer";
-import { ScopeSupport } from "@cursorless/cursorless-engine";
 
 export class VscodeIterationScopeVisualizer extends VscodeScopeVisualizer {
   protected getScopeSupport(editor: TextEditor): ScopeSupport {
